requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0
networkx>=2.6.0
torch>=1.10.0
PyYAML>=5.4.0
matplotlib>=3.4.0
seaborn>=0.11.0
PyYAML>=5.4.0
matplotlib>=3.4.0
seaborn>=0.11.0
flask>=2.0.0
flask-cors>=3.0.10
